# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/map/points": {
      "get": {
        "operationId": "MapController_points",
        "summary": "Marcadores de datasets para el mapa (uno por dataset), con su info/summary",
        "parameters": [
          {
            "name": "ocean",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "mediterraneo",
                "atlantico",
                "catambrico"
              ],
              "type": "string"
            }
          },
          {
            "name": "datasetType",
            "required": false,
            "in": "query",
            "description": "p. ej. recogidas_playa, boya_biomasa_slx+",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "geojson → FeatureCollection [lng,lat]",
            "schema": {
              "enum": [
                "json",
                "geojson"
              ],
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "Con token: 'mine' (por defecto) limita a tu organización",
            "schema": {
              "enum": [
                "mine",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Map"
        ]
      }
    },
...
    "/v1/admin/organizations": {
      "get": {
        "operationId": "AdminController_listOrganizations",
        "summary": "List data space participants",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_createOrganization",
        "summary": "Create a data space participant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateOrganization",
        "summary": "Update a participant (matched by slug)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "summary": "List users",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_createUser",
        "summary": "Create a user attached to a participant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/admin/users/password": {
      "put": {
        "operationId": "AdminController_setPassword",
        "summary": "Reset a user's password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Log in with email + password",
        "description": "Returns a JWT for the Authorization: Bearer header. The `username` field also accepts a legacy portal connector username while those accounts are being migrated.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalLoginResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Portal auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Identity behind the current token",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Portal auth"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/sync/assets": {
      "post": {
        "operationId": "SyncController_syncAsset",
        "summary": "Ingest one asset that was uploaded or updated in the data space",
        "description": "Fetches the object, validates its container, normalizes it and replaces its observations in one transaction. Idempotent: an unchanged checksum is reported as \"unchanged\" without writing.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAssetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such object in the bucket"
          },
          "422": {
            "description": "The object exists but is not a usable asset"
          }
        },
        "tags": [
          "Dataspace sync"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/sync/scan": {
      "post": {
        "operationId": "SyncController_scan",
        "summary": "Reconcile a prefix against the bucket",
        "description": "Ingests new and changed objects, leaves unchanged ones alone, and flags assets whose object has disappeared (their observations are kept).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncScanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Dataspace sync"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/sync/runs": {
      "get": {
        "operationId": "SyncController_listRuns",
        "summary": "Sync history",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dataspace sync"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    },
    "/v1/sync/runs/{id}": {
      "get": {
        "operationId": "SyncController_getRun",
        "summary": "One sync run, with its per-asset results",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dataspace sync"
        ],
        "security": [
          {
            "portal-jwt": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Analytics API of ONDAs DataSpace",
    "description": "Documentación del API de analíticas. Los nombres de los campos se mantienen en inglés. POST /v1/auth/login devuelve JWT; usar cabecera Authorization: Bearer (token) en POST /v1/analyses/run.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "portal-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token de POST /v1/auth/login"
      }
    },
    "schemas": {
...
      "MapPointDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "atlantico/innoceana/recogidas_playa_tenerife"
          },
          "name": {
            "type": "string",
            "example": "Tenerife — Coastal cleanup"
          },
          "datasetType": {
            "type": "string",
            "example": "recogidas_playa"
          },
          "label": {
            "type": "string",
            "example": "Coastal cleanup"
          },
          "category": {
            "type": "string",
            "enum": [
              "cleanup",
              "biomass",
              "microplastics",
              "environmental",
              "atmospheric",
              "oceanographic"
            ]
          },
          "color": {
            "type": "string",
            "example": "#00003F"
          },
          "provider": {
            "type": "string",
            "example": "innoceana"
          },
          "ocean": {
            "type": "string",
            "example": "atlantico"
          },
          "lat": {
            "type": "number",
            "example": 28.1876
          },
          "lng": {
            "type": "number",
            "example": -16.6596
          },
          "records": {
            "type": "object",
            "example": 8,
            "nullable": true
          },
          "dateRange": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DateRangeDto"
              }
            ]
          },
          "format": {
            "type": "string",
            "example": "rows"
          },
          "units": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "Biomass depth -3_-5 m": "Tonnes"
            }
          },
          "url": {
            "type": "string",
            "example": "https://universalplastic-sedia.s3.eu-central-1.amazonaws.com/public/atlantico/innoceana/recogidas_playa_tenerife.json"
          },
          "metadataSchemaRef": {
            "type": "object",
            "nullable": true
          },
          "warnings": {
            "example": [
              "coords corrected 31.483,-11.926 → Tenerife (28.188,-16.660)"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "object",
            "additionalProperties": true,
            "description": "Type-specific headline stats",
            "example": {
              "kg": 144.87,
              "volunteers": 31,
              "cleanups": 8
            }
          },
          "cleanupsList": {
            "description": "cleanup datasets only",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CleanupEventDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "datasetType",
          "label",
          "category",
          "color",
          "provider",
          "ocean",
          "lat",
          "lng",
          "records",
          "format",
          "url",
          "metadataSchemaRef",
          "warnings"
        ]
      },
      "MapResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 22
          },
          "bounds": {
            "type": "object",
            "example": [
              [
                28.1876,
                -16.6596
              ],
              [
                43.5721,
                2.795
              ]
            ],
            "description": "[[minLat,minLng],[maxLat,maxLng]] for map.fitBounds",
            "nullable": true
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapPointDto"
            }
          }
        },
        "required": [
          "count",
          "bounds",
          "points"
        ]
      },
      "OrganizationS3Dto": {
        "type": "object",
        "properties": {
          "bucket": {
            "type": "string",
            "example": "universalplastic-sedia"
          },
          "prefix": {
            "type": "string",
            "example": "public/mediterraneo/universal_plastic/",
            "description": "This participant's space in the data space bucket"
          }
        }
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "innoceana",
            "description": "Stable slug, lowercase, no spaces"
          },
          "name": {
            "type": "string",
            "example": "Innoceana"
          },
          "type": {
            "type": "string",
            "example": "NGO",
            "enum": [
              "Company",
              "NGO",
              "Institution",
              "Campaign"
            ]
          },
          "territory": {
            "type": "string",
            "example": "Costa Brava, Catalunya, Spain"
          },
          "description": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "example": "innoceana.org"
          },
          "contact": {
            "type": "string",
            "example": "impact@innoceana.org"
          },
          "publicProfile": {
            "type": "boolean",
            "default": true
          },
          "dataProviderIds": {
            "example": [
              "innoceana"
            ],
            "description": "Every spelling of this org dataProviderId found in its files",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "providerFolders": {
            "example": [
              "innoceana"
            ],
            "description": "S3 provider folders owned by this org (public/{ocean}/{folder}/)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "s3": {
            "$ref": "#/components/schemas/OrganizationS3Dto"
          }
        },
        "required": [
          "slug",
          "name"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "impact@innoceana.org"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "description": "At least 10 characters"
          },
          "name": {
            "type": "string",
            "example": "Innoceana Impact Team"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "provider",
              "viewer"
            ],
            "example": "provider"
          },
          "organizationSlug": {
            "type": "string",
            "example": "innoceana",
            "description": "Required for role \"provider\""
          },
          "legacyUsername": {
            "type": "string",
            "example": "user_innoceana",
            "description": "Legacy portal connector username"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "impact@innoceana.org"
          },
          "password": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PortalLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "impact@universalplastic.io",
            "description": "Email of the user. Legacy portal connector usernames are still accepted."
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "object",
            "nullable": true
          },
          "name": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "provider",
              "viewer"
            ]
          },
          "organization": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "email",
          "name",
          "role",
          "organization"
        ]
      },
      "PortalLoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "number",
            "description": "Token TTL in seconds"
          },
          "username": {
            "type": "string",
            "description": "Email, or the legacy connector username"
          },
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            ]
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "username",
          "user"
        ]
      },
      "SyncAssetDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "public/mediterraneo/port_badalona/boya_biomasa_badalona.json",
            "description": "S3 object key of the asset that was uploaded or updated"
          },
          "url": {
            "type": "string",
            "example": "https://universalplastic-sedia.s3.eu-central-1.amazonaws.com/public/…/file.json",
            "description": "Full object URL — an alternative to `key`"
          },
          "force": {
            "type": "boolean",
            "default": false,
            "description": "Re-ingest even when the object checksum is unchanged"
          }
        }
      },
...
      "SyncRunResponseDto": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "asset",
              "scan"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "ok",
              "partial",
              "failed"
            ]
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "finishedAt": {
            "type": "object",
            "nullable": true
          },
          "totals": {
            "type": "object",
            "description": "assets/created/updated/unchanged/missing/failed/observations/warnings"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncResultRowDto"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "runId",
          "kind",
          "status",
          "startedAt",
          "finishedAt",
          "totals",
          "results",
          "warnings"
        ]
      },
      "SyncScanDto": {
        "type": "object",
        "properties": {
          "prefix": {
            "type": "string",
            "example": "public/mediterraneo/",
            "description": "Prefix to reconcile. Defaults to your organization space, or the whole root for admins."
          },
          "dryRun": {
            "type": "boolean",
            "default": false,
            "description": "Report the plan without writing anything"
          },
          "force": {
            "type": "boolean",
            "default": false,
            "description": "Re-ingest every object, changed or not"
          }
        }
      }
    }
  }
}